// lib/requests/types.gen.ts
export type models_Order = {
    createdAt?: string;
    deletedAt?: string;
    id?: string;
    products: Array<models_Product>;
    status: string;
    total?: number;
    tracking: string;
    updatedAt?: string;
    userID?: string;
};

export type models_Product = {
    id?: string;
    image: string;
    name: string;
    price: number;
    quantity: number;
    size: string;
};